requirements.txt
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.2.0
tensorflow==2.4.1